905: Expose WfoMenuLinkItem component for consumer who want to expand the sidebarMenu
