+++ packages/orchestrator-ui-components/src/components/WfoPageTemplate/WfoSidebar/index.ts
@@ -1,1 +1,2 @@
 export * from './WfoSidebar';
+export * from './WfoMenuLink';
